newchat-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.47.0
google-generativeai==0.7.2
pydantic==2.9.2
httpx==0.27.2
langchain==0.3.1
langgraph==0.2.15

